API

ClimaRecursiveApply.ClimaRecursiveApply — Module

ClimaRecursiveApply

This module contains operators to recurse over nested Tuples or NamedTuples.

To extend to another type T, define ClimaRecursiveApply.rmap(fn, args::T...)

Map

ClimaRecursiveApply.rmap — Function

rmap(fn, X...)

Recursively apply fn to each element of X

ClimaRecursiveApply.rmaptype — Function

rmaptype(fn, T)
rmaptype(fn, T1, T2)

Recursively apply fn to each type parameter of the type T, or to each type parameter of the types T1 and T2, where fn returns a type.

Conversions

ClimaRecursiveApply.rconvert — Function

rconvert(T, X)

Identical to convert(T, X), but with improved type stability for nested types.

ClimaRecursiveApply.rpromote_type — Function

rpromote_type(Ts...)

Recursively apply promote_type to the input types.

Math

ClimaRecursiveApply.rzero — Function

rzero(T)

Recursively compute the zero value of type T.

ClimaRecursiveApply.rmul — Function

rmul(X, Y)
X ⊠ Y

Recursively scale each element of X by Y.

ClimaRecursiveApply.radd — Function

radd(X, Y)
X ⊞ Y

Recursively add elements of X and Y.

ClimaRecursiveApply.rsub — Function

rsub(X, Y)
X ⊟ Y

Recursively subtract elements of Y from X.

ClimaRecursiveApply.rdiv — Function

rdiv(X, Y)

Recursively divide each element of X by Y

ClimaRecursiveApply.rmuladd — Function

rmuladd(w, X, Y)

Recursively add elements of w * X + Y.

ClimaRecursiveApply.rmatmul1 — Function

rmatmul1(W, S, i, j)

Recursive matrix product along the 1st dimension of S. Equivalent to:

mapreduce(⊠, ⊞, W[i,:], S[:,j])

ClimaRecursiveApply.rmatmul2 — Function

rmatmul2(W, S, i, j)

Recursive matrix product along the 2nd dimension S. Equivalent to:

mapreduce(⊠, ⊞, W[j,:], S[i, :])

ClimaRecursiveApply.rmin — Function

rmin(x, y)

Recursively apply min

ClimaRecursiveApply.rmax — Function

rmax(x, y)

Recursively apply max